requirements.txt
# Core Framework Dependencies
langgraph==0.4.8
langchain==0.3.25
langchain-core==0.3.64
langchain-community==0.3.24

# AI Model Integrations
openai==1.84.0
langchain-openai==0.3.21
langchain-anthropic==0.3.15
langchain-google-genai==2.1.5
langchain-nvidia-ai-endpoints==0.3.10

# LangGraph Extensions
langgraph-swarm==0.0.11
langgraph-cli==0.2.12
langgraph-checkpoint==2.0.26
langgraph-checkpoint-postgres==2.0.21
langgraph-checkpoint-sqlite==2.0.10
langgraph-checkpoint-mongodb==0.1.3

# Database Drivers
psycopg[binary,pool]==3.2.9
asyncpg==0.30.0
pymongo==4.11.3

# Web Framework
fastapi==0.115.12
uvicorn==0.34.3
streamlit==1.45.1

# Core Libraries
pydantic==2.11.5
pydantic-settings==2.9.1
python-dotenv==1.1.0
httpx==0.28.1
requests==2.32.3

# Data Processing
pandas==2.3.0
numpy==2.3.0

# Async Support
nest-asyncio==1.6.0

# Search and Tools
duckduckgo-search==8.0.2

# Utilities
tiktoken==0.9.0
python-multipart==0.0.20

# Monitoring and Observability
langsmith==0.3.45